Add support for loops in the NASL Rust implementation
This PR introduces a new module which adds support to interprete loop structures in the rust implementation of NASL.
In order to support the loop control flow keywords `continue` and `break` the nasl-syntax was adjusted to parse them.
To be able to actually work with these controll flow keywords, new `NaslValue`s are introduced.